+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Django 3.1
 - Handle status callbacks from Twilio, show delivery failures in Integrations
 - Removing unused /api/v1/notifications/{uuid}/bounce endpoint
+- Less verbose output in the `senddeletionnotices` command
 
 ## Bug Fixes
 - Handle excessively long email addresses in the signup form.